# api/docs/swagger.yaml
definitions:
...
  models.CreateEmployee:
    properties:
      address:
        type: string
      department_id:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      role_id:
        type: string
    type: object
  models.CreateRole:
    properties:
      discription:
        type: string
      r_title:
        type: string
    type: object
...
  models.Employee:
    properties:
      address:
        type: string
      created_at:
        type: string
      department_id:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      role_id:
        type: string
      updated_at:
        type: string
    type: object
  models.Empty:
    type: object
...
  models.GetListEmployeeResponse:
    properties:
      count:
        type: integer
      employees:
        items:
          $ref: '#/definitions/models.Employee'
        type: array
    type: object
  models.GetListRoleResponse:
    properties:
      count:
        type: integer
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
    type: object
...
  models.Role:
    properties:
      created_at:
        type: string
      discription:
        type: string
      id:
        type: string
      r_title:
        type: string
      updated_at:
        type: string
    type: object
...
  models.UpdateEmployeeSwag:
    properties:
      address:
        type: string
      department_id:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      role_id:
        type: string
    type: object
  models.UpdateRoleSwag:
    properties:
      discription:
        type: string
      r_title:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /employee:
    get:
      consumes:
      - application/json
      description: Get List Employee
      operationId: get_list_employee
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: GetEmployeeListBody
          schema:
            $ref: '#/definitions/models.GetListEmployeeResponse'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get List Employee
      tags:
      - Employee
    post:
      consumes:
      - application/json
      description: Create Employee
      operationId: create_employee
      parameters:
      - description: CreateEmployeeRequestBody
        in: body
        name: Employee
        required: true
        schema:
          $ref: '#/definitions/models.CreateEmployee'
      produces:
      - application/json
      responses:
        "201":
          description: GetEmployeeBody
          schema:
            $ref: '#/definitions/models.Employee'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Create Employee
      tags:
      - Employee
  /employee/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Employee
      operationId: delete_employee
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: DeleteEmployeeBody
          schema:
            $ref: '#/definitions/models.Empty'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Delete Employee
      tags:
      - Employee
    get:
      consumes:
      - application/json
      description: Get By ID Employee
      operationId: get_by_id_employee
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetEmployeeBody
          schema:
            $ref: '#/definitions/models.Employee'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get By ID Employee
      tags:
      - Employee
    put:
      consumes:
      - application/json
      description: Update Employee
      operationId: update_employee
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateEmployeeRequestBody
        in: body
        name: Employee
        required: true
        schema:
          $ref: '#/definitions/models.UpdateEmployeeSwag'
      produces:
      - application/json
      responses:
        "202":
          description: UpdateEmployeeBody
          schema:
            $ref: '#/definitions/models.Employee'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Update Employee
      tags:
      - Employee
...
  /role:
    get:
      consumes:
      - application/json
      description: Get List Role
      operationId: get_list_Role
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetRoleListBody
          schema:
            $ref: '#/definitions/models.GetListRoleResponse'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get List Role
      tags:
      - Role
    post:
      consumes:
      - application/json
      description: Create Role
      operationId: create_Role
      parameters:
      - description: CreateRoleRequestBody
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.CreateRole'
      produces:
      - application/json
      responses:
        "201":
          description: GetRoleBody
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Create Role
      tags:
      - Role
  /role/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Role
      operationId: delete_Role
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: DeleteRoleBody
          schema:
            $ref: '#/definitions/models.Empty'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Delete Role
      tags:
      - Role
    get:
      consumes:
      - application/json
      description: Get By ID Role
      operationId: get_by_id_Role
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetRoleBody
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get By ID Role
      tags:
      - Role
    put:
      consumes:
      - application/json
      description: Update Role
      operationId: update_Role
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateRoleRequestBody
        in: body
        name: Role
        required: true
        schema:
          $ref: '#/definitions/models.UpdateRoleSwag'
      produces:
      - application/json
      responses:
        "202":
          description: UpdateRoleBody
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Update Role
      tags:
      - Role
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"